+++ circuit_maintenance_parser/parsers/verizon.py
@@ -5,7 +5,7 @@ from typing import Dict
 from dateutil import parser
 from bs4.element import ResultSet  # type: ignore
 
-from circuit_maintenance_parser.errors import ParsingError
+from circuit_maintenance_parser.errors import ParserError
 from circuit_maintenance_parser.parser import Html, Impact, CircuitImpact, Status
 
 logger = logging.getLogger(__name__)
@@ -26,7 +26,7 @@ class HtmlParserVerizon1(Html):
             return [data]
 
         except Exception as exc:
-            raise ParsingError from exc
+            raise ParserError from exc
 
     def parse_tables(self, tables: ResultSet, data: Dict):  # pylint: disable=too-many-locals
         """Parse <table> tag."""